add parts of broadcast

+++ src/main/java/com/example/radio/repositories/BroadcastRepository.java
@@ -3,5 +3,5 @@ package com.example.radio.repositories;
 import com.example.radio.entities.Broadcast;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BroadcastRepository extends JpaRepository<Broadcast, Long> {
+public interface BroadcastRepository extends JpaRepository<Broadcast, Integer> {
 }
